refactor: move fitCameraToBoundingBox as a part of utilities
