+++ viewer/packages/camera-manager/src/CameraManagerHelper.ts
@@ -85,29 +85,6 @@ export class CameraManagerHelper {
     camera.far = far;
     camera.updateProjectionMatrix();
   }
-  /**
-   * Calculates camera position and target that allows to see the content of provided bounding box.
-   * @param camera Used camera instance.
-   * @param box Bounding box to be fitted.
-   * @param radiusFactor The ratio of the distance from camera to center of box and radius of the box.
-   * @returns
-   */
-  static calculateCameraStateToFitBoundingBox(
-    camera: THREE.PerspectiveCamera,
-    box: THREE.Box3,
-    radiusFactor: number = 2
-  ): { position: THREE.Vector3; target: THREE.Vector3 } {
-    const boundingSphere = box.getBoundingSphere(new THREE.Sphere());
-
-    const target = boundingSphere.center;
-    const distance = boundingSphere.radius * radiusFactor;
-    const direction = new THREE.Vector3(0, 0, -1);
-    direction.applyQuaternion(camera.quaternion);
-
-    const position = direction.clone().multiplyScalar(-distance).add(target);
-
-    return { position, target };
-  }
 
   private static calculateCameraFar(
     near: number,
